Implement disemvowel
This commit adds the necessary code to implement disemvoweling the input string and writing that disemovoweled string to the output file specified in the arguments. Right now, most of the tests pass, but the tests regarding the panic! at requiring two arugments and the test about panic! when you cannot read the input file are failing.